[Refactor] 다이얼로그,바텀시트 리팩토링 및 어댑터 아이디 이슈 수정
