+++ app/src/main/java/com/hara/kaera/presentation/write/custom/DialogSaveWarning.kt
@@ -1,4 +1,4 @@
-package com.hara.kaera.presentation.write.dialog
+package com.hara.kaera.presentation.write.custom
 
 import android.os.Bundle
 import android.view.View
